+++ test/duration_test.exs
@@ -16,6 +16,18 @@ defmodule Moar.DurationTest do
     end
   end
 
+  describe "to_short_string" do
+    test "converts a duration to a short (non-localized) string" do
+      assert Moar.Duration.to_short_string({25, :day}) == "25d"
+      assert Moar.Duration.to_short_string({25, :hour}) == "25h"
+      assert Moar.Duration.to_short_string({25, :minute}) == "25m"
+      assert Moar.Duration.to_short_string({25, :second}) == "25s"
+      assert Moar.Duration.to_short_string({25, :millisecond}) == "25ms"
+      assert Moar.Duration.to_short_string({25, :microsecond}) == "25us"
+      assert Moar.Duration.to_short_string({25, :nanosecond}) == "25ns"
+    end
+  end
+
   describe "to_string" do
     test "converts a duration to a (non-localized) string" do
       assert Moar.Duration.to_string({1, :second}) == "1 second"